% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_knapsack.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Brute force search. A function that takes a data frame and returns the maximum knapsack value.}
\usage{
knapsack_brute_force(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{numeric value}
}
\value{
m list
}
\description{
Brute force search. A function that takes a data frame and returns the maximum knapsack value.
}

