% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming algorithm for the knapsack problem}
\usage{
knapsack_dynamic(X, W)
}
\arguments{
\item{W}{is the total size of the knapsack}

\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}
}
\description{
Dynamic programming algorithm for the knapsack problem
}
\details{
An algorithm that can solve the knapsack problem exact by iterating over all possible values of w all possbile values with good accuracy , and also gives the maximum value for the knapsack
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
}
\author{
Aqeel Ahmed, Marhawi Tewolde
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\seealso{
\code{\link{Dynamic Programming Knapsack}}
}

