% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_knapsack.R
\name{knapsack_dynamic_optimized}
\alias{knapsack_dynamic_optimized}
\title{Dynamic programming (Optimized version, faster)}
\usage{
knapsack_dynamic_optimized(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{numeric value}
}
\value{
out list
}
\description{
Dynamic programming (Optimized version, faster)
}

