% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_knapsack.R
\name{knapsack_greedy}
\alias{knapsack_greedy}
\title{Greedy heuristic. This algorithm gives an approximation of the result, which will reduce the computional complexity considerably.
Complexity analysis:
	- For calculate value/weight: n
 - For sorting: fastest algorithm is O(nlogn)
	- For select element into list: n
=> Complexity: O(nlog) for n = 1000000: 6e+6}
\usage{
knapsack_greedy(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{numeric value}
}
\value{
max numeric list
}
\description{
Greedy heuristic. This algorithm gives an approximation of the result, which will reduce the computional complexity considerably.
Complexity analysis:
	- For calculate value/weight: n
 - For sorting: fastest algorithm is O(nlogn)
	- For select element into list: n
=> Complexity: O(nlog) for n = 1000000: 6e+6
}

