\name{KnapsackPackage-package}
\alias{KnapsackPackage-package}
\alias{KnapsackPackage}
\docType{package}
\title{
\packageTitle{KnapsackPackage}
}
\description{
\packageDescription{KnapsackPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KnapsackPackage}
\packageIndices{KnapsackPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{KnapsackPackage}

Maintainer: \packageMaintainer{KnapsackPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
