% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagedoc.R
\docType{package}
\name{knapsackpackage}
\alias{knapsackpackage}
\alias{knapsackpackage-package}
\title{knapsackpackage: 
                 
                 
                 
The knapsack package provides three function.A package for computating the knapsack problem in three differnt ways
                     1, Brute force
                     2. Dynamic
                     3. Greedy}
\description{
knapsackpackage: 
                 
                 
                 
The knapsack package provides three function.A package for computating the knapsack problem in three differnt ways
                     1, Brute force
                     2. Dynamic
                     3. Greedy
}
\section{knapsackpackage functions}{
they took dataframe and one scalar as a limit weifht
                                   and all of them return the same result with the optimum
                                   value with the weight of the item is not greater than the given weight.
}

