% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic.R
\name{seq_of_decision}
\alias{seq_of_decision}
\title{Helper Method To Get The Correct Indexes}
\usage{
seq_of_decision(final_mat, val)
}
\arguments{
\item{final_mat}{Final matrix of values}

\item{val}{Final value}
}
\value{
A list containing two elements: value which contains the sum of the
elements and elements containing the set of elements accommodated in the 
knapsack
}
\description{
Helper Method To Get The Correct Indexes
}
