% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack_brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solving Knapsack Problem By Brute Force Method}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data frame containing elements to be added to the knapsack}

\item{W}{Maximum weight of the knapsack}

\item{parallel}{Argument to parallelize brute force}
}
\value{
A list containing two elements: value which contains the sum of the
elements and elements containing the set of elements accommodated in the 
knapsack
}
\description{
Solving Knapsack Problem By Brute Force Method
}
