% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic Approach For Knapsack Problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data frame containing elements to be added to the knapsack}

\item{W}{Maximum weight of the knapsack}
}
\value{
A list containing two elements: value which contains the sum of the
elements and elements containing the set of elements accommodated in the 
knapsack
}
\description{
Greedy Heuristic Approach For Knapsack Problem
}
