% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic. This algorithm gives an approximation of the result, which will reduce the computional complexity considerably.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{numeric value}
}
\value{
max numeric list
}
\description{
Greedy heuristic. This algorithm gives an approximation of the result, which will reduce the computional complexity considerably.
}

