% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solving Knapsack Problem By Dynamic Method}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Data frame containing elements to be added to the knapsack}

\item{W}{Maximum weight of the knapsack}
}
\value{
A list containing two elements: value which contains the sum of the
elements and elements containing the set of elements accommodated in the 
knapsack
}
\description{
Solving Knapsack Problem By Dynamic Method
}
