% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming. Creates an algorithm that can solve the knapsack problem exact by iterating over all possible values of w.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame}

\item{W}{numeric value}
}
\value{
out list
}
\description{
Dynamic programming. Creates an algorithm that can solve the knapsack problem exact by iterating over all possible values of w.
}

