% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic.R
\name{seq_of_decision}
\alias{seq_of_decision}
\title{Helper Method To Get The Correct Indexes}
\usage{
seq_of_decision(final_mat, val)
}
\arguments{
\item{final_mat}{Calculated matrix}

\item{val}{Value to be found in the matrix rows}
}
\value{
The index of the row containing the value
}
\description{
Helper Method To Get The Correct Indexes
}
